Support implicit `count` with opts
`first/2` and `last/2` can be called with unexpected args, e.g., `UserSingle.first(order_by: :info)`. Once might expect that to return the 1st record ordered by info ascending. Instead, it returns a query like "from u0 in UserSingle, order_by: [asc: u0.id], limit: ^[{:order_by, :info}]".
Should we consider checking to see if count is an integer and defaulting it to 1 if it's not?